make markdown writer a separate file in writers folder
